*refactoring of ConvertWhereNode.
